backend_server/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==6.0.1
sentence-transformers==5.1.0
chromadb==1.0.16